Restructure codebase
... to clearly separate direct and indirect clickers.
