+++ src/background/postinstall.ts
