+++ src/indirectClickers/cookiebot.ts
@@ -1,4 +1,4 @@
-import { clickAllWhenFound, clickWhenFound } from "./common";
+import { clickAllWhenFound, clickWhenFound } from "../common";
 
 async function handle() {
     await clickAllWhenFound('input.CybotCookiebotDialogBodyLevelButton:checked:not(:disabled)');
